ListView and GridView support
This adds support for finding objects in ListViews or GridViews.
Spix will now search for objects as `children()` of QObjects,
but also as `childItems()` of QQuickItems. With this, items in
ListViews and GridViews are found via ItemPaths.
This also means that it is no longer required to explicitly
specify a repeater object in the path when accessing its
children.