introduce ItemPathComponent class to handle name and property selectors
Rather than using a string to represent the component,
we now use a class that can handle both name and property selectors.
This allows us to add more complex selectors in the future without
having to change the ItemPath class.
It also moves the logic of parsing the component string from the scene
to the ItemPathComponent class.