Introduce Path Selectors (#136)
Rather than using a string to represent the components of an ItemPath,
we now use a class that can handle both name and property selectors.
These selectors can then be used by a Scene to determine if a certain object
matches that selector / path component.
This also moves the logic of parsing the component string from the scene
to the ItemPathComponent class.