Fix item lookup (#142)
Since version 0.11 QtScene.itemAtPath always returns an Item filled with a QQuickItem. If you want to access properties of the mainWindow, this is not working, if you want e.g. change the main window size.

This fix returns a QQuickWindow if the lookup path has just one element. If there are more elements, a QQuickItem is returned (the current behaviour for all cases).

---------

Co-authored-by: Falko Axmann <code@falkoaxmann.de>